Move ZipModLoader into it's own file
